fix problem on start of new bash after all closed
previously after closing all bash  windows then loading a new one the ssh agent or something caused a red error message and the terminal to be useless.

Now it only starts and loads the keys on the first term opened since system startup.  If all are closed and a new is opened it uses the existing ssh-agent.